Upgrade to .NET 9.0 and add FribergCarRentals.Data
Updated the solution to target .NET 9.0, including upgrading
`Microsoft.EntityFrameworkCore` packages to version 9.0.10
and `Microsoft.VisualStudio.Web.CodeGeneration.Design` to
version 9.0.0.

Added a new project, `FribergCarRentals.Data`, to the solution
with build configurations and support for `ImplicitUsings`
and `Nullable` features. Updated the solution file to include
the new project and its configurations.